ci: add local-ci suite and fix workspace clippy lints
- Add ./scripts/local-ci.sh script for full workspace validation
- Fix workspace Clippy warnings under -D warnings across consensus, network, rpc, mempool, sdk, storage, and node crates
- Add insecure-password-verify feature to crates/crypto/Cargo.toml
- Update README.md with Local CI verification badge and status table entry

+++ crates/consensus/src/asert.rs
@@ -636,7 +636,7 @@ mod tests {
         let slow_int = asert_next_target(
             compact_to_target(anchor_bits),
             window as i64,
-            slow_delta as i64,
+            slow_delta,
             &params,
             None,
         );
